requirements.txt
django-widget-tweaks
uuid